refactor: modularize ebpf udp ingress path
