+++ dataplane/ebpf/src/main.rs
@@ -1,65 +1,39 @@
 #![no_std]
 #![no_main]
 
-use core::mem;
-mod ingress;
-
-use aya_bpf::{
-    bindings::{TC_ACT_OK, TC_ACT_PIPE, TC_ACT_SHOT},
-    helpers::{bpf_csum_diff, bpf_redirect_neigh},
-    macros::{classifier, map},
-    maps::HashMap,
-    programs::TcContext,
-};
-use aya_log_ebpf::info;
-use mem::size_of;
-use memoffset::offset_of;
-
-use common::{Backend, BackendKey};
-
 #[allow(non_upper_case_globals)]
 #[allow(non_snake_case)]
 #[allow(non_camel_case_types)]
 #[allow(dead_code)]
 mod bindings;
+mod ingress;
+mod utils;
 
-use bindings::{ethhdr, iphdr, udphdr};
+use memoffset::offset_of;
 
-const ETH_P_IP: u16 = 0x0800;
+use aya_bpf::{
+    bindings::{TC_ACT_OK, TC_ACT_PIPE, TC_ACT_SHOT},
+    macros::{classifier, map},
+    maps::HashMap,
+    programs::TcContext,
+};
 
-const IPPROTO_UDP: u8 = 17;
+use bindings::{ethhdr, iphdr};
+use common::{Backend, BackendKey};
+use ingress::{tcp::handle_tcp_ingress, udp::handle_udp_ingress};
+use utils::{ETH_HDR_LEN, ETH_P_IP, IPPROTO_TCP, IPPROTO_UDP};
 
-const ETH_HDR_LEN: usize = mem::size_of::<ethhdr>();
-const IP_HDR_LEN: usize = mem::size_of::<iphdr>();
-
-fn ip_from_int(ip: u32) -> [u8; 4] {
-    let mut addr: [u8; 4] = [0; 4];
-
-    addr[0] = ((ip >> 0) & 0xFF) as u8;
-    addr[1] = ((ip >> 8) & 0xFF) as u8;
-    addr[2] = ((ip >> 16) & 0xFF) as u8;
-    addr[3] = ((ip >> 24) & 0xFF) as u8;
-
-    addr
-}
-
-// Gives us raw pointers to a specific offset in the packet
-#[inline(always)]
-unsafe fn ptr_at<T>(ctx: &TcContext, offset: usize) -> Result<*mut T, i64> {
-    let start = ctx.data();
-    let end = ctx.data_end();
-    let len = mem::size_of::<T>();
-
-    if start + offset + len > end {
-        return Err(TC_ACT_OK.into());
-    }
-
-    Ok((start + offset) as *mut T)
-}
+// -----------------------------------------------------------------------------
+// Maps
+// -----------------------------------------------------------------------------
 
 #[map(name = "BACKENDS")]
 static mut BACKENDS: HashMap<BackendKey, Backend> =
     HashMap::<BackendKey, Backend>::with_max_entries(128, 0);
+
+// -----------------------------------------------------------------------------
+// Ingress
+// -----------------------------------------------------------------------------
 
 #[classifier(name = "tc_ingress")]
 pub fn tc_ingress(ctx: TcContext) -> i32 {
@@ -69,19 +43,6 @@ pub fn tc_ingress(ctx: TcContext) -> i32 {
     };
 
     return TC_ACT_OK;
-}
-
-fn get_backend(key: BackendKey) -> Option<&'static Backend> {
-    unsafe { BACKENDS.get(&key) }
-}
-
-fn csum_fold_helper(mut csum: u64) -> u16 {
-    for _i in 0..4 {
-        if (csum >> 16) > 0 {
-            csum = (csum & 0xffff) + (csum >> 16);
-        }
-    }
-    return !(csum as u16);
 }
 
 // Make sure ip_forwarding is enabled on the interface this it attached to
@@ -99,76 +60,35 @@ fn try_tc_ingress(ctx: TcContext) -> Result<i32, i64> {
         .load::<u8>(ETH_HDR_LEN + offset_of!(iphdr, protocol))
         .map_err(|_| TC_ACT_PIPE)?;
 
-    if protocol != IPPROTO_UDP {
-        return Ok(TC_ACT_PIPE);
+    match protocol {
+        IPPROTO_TCP => handle_tcp_ingress(ctx),
+        IPPROTO_UDP => handle_udp_ingress(ctx),
+        _ => Ok(TC_ACT_PIPE),
     }
-
-    let ip_hdr: *mut iphdr = unsafe { ptr_at(&ctx, ETH_HDR_LEN) }?;
-
-    let udp_header_offset = ETH_HDR_LEN + IP_HDR_LEN;
-
-    let udp_hdr: *mut udphdr = unsafe { ptr_at(&ctx, udp_header_offset)? };
-
-    let key = BackendKey {
-        ip: u32::from_be(unsafe { (*ip_hdr).daddr }),
-        port: (u16::from_be(unsafe { (*udp_hdr).dest })) as u32,
-    };
-
-    let backend = get_backend(key).ok_or(TC_ACT_OK)?;
-
-    let daddr_dot_dec = ip_from_int(unsafe { (*ip_hdr).daddr });
-    info!(
-        &ctx,
-        "Received a packet destined for svc ip: {}.{}.{}.{} at port: {}",
-        daddr_dot_dec[0],
-        daddr_dot_dec[1],
-        daddr_dot_dec[2],
-        daddr_dot_dec[3],
-        u16::from_be(unsafe { (*udp_hdr).dest })
-    );
-
-    // Update destination IP
-    unsafe {
-        (*ip_hdr).daddr = backend.daddr.to_be();
-    }
-
-    if (ctx.data() + ETH_HDR_LEN + size_of::<iphdr>()) > ctx.data_end() {
-        info!(&ctx, "Iphdr is out of bounds");
-        return Ok(TC_ACT_OK);
-    }
-
-    // Calculate l3 cksum
-    // TODO(astoycos) use l3_cksum_replace instead
-    unsafe { (*ip_hdr).check = 0 };
-    let full_cksum = unsafe {
-        bpf_csum_diff(
-            mem::MaybeUninit::zeroed().assume_init(),
-            0,
-            ip_hdr as *mut u32,
-            size_of::<iphdr>() as u32,
-            0,
-        )
-    } as u64;
-    unsafe { (*ip_hdr).check = csum_fold_helper(full_cksum) };
-
-    // Update destination port
-    unsafe { (*udp_hdr).dest = (backend.dport as u16).to_be() };
-    // Kernel allows UDP packet with unset checksums
-    unsafe { (*udp_hdr).check = 0 };
-
-    let action = unsafe {
-        bpf_redirect_neigh(
-            backend.ifindex as u32,
-            mem::MaybeUninit::zeroed().assume_init(),
-            0,
-            0,
-        )
-    };
-
-    info!(&ctx, "redirect action: {}", action);
-
-    Ok(action as i32)
 }
+
+// -----------------------------------------------------------------------------
+// Egress
+// -----------------------------------------------------------------------------
+
+#[classifier(name = "tc_egress")]
+pub fn tc_egress(ctx: TcContext) -> i32 {
+    match try_tc_egress(ctx) {
+        Ok(ret) => ret,
+        Err(_) => TC_ACT_SHOT,
+    };
+
+    return TC_ACT_OK;
+}
+
+fn try_tc_egress(_ctx: TcContext) -> Result<i32, i64> {
+    // TODO: not implemented yet
+    Ok(TC_ACT_PIPE)
+}
+
+// -----------------------------------------------------------------------------
+// Panic Implementation
+// -----------------------------------------------------------------------------
 
 #[panic_handler]
 fn panic(_info: &core::panic::PanicInfo) -> ! {
